requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
PyYAML==6.0.2
